Create Clenow Momentum Strategy (as amended)
This was cloned from a lineage of discourse in the Quantopian community, attempting to code the ideas from Andreas Clenow's book (I think he wrote the first Algo).  https://www.quantopian.com/posts/stocks-on-the-move-by-andreas-clenow.  I started with the latest version of the algo, that has been substantially iterated upon, by Charles Witt, Burrito Dan, and others.